backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
authlib>=1.2.0
bcrypt>=4.0.0
itsdangerous>=2.1.0
starlette>=0.27.0
duckdb>=1.4.0
chardet>=5.0.0
